requirements.txt
Django~=3.1.5
django-sendgrid-v5
Pillow
django-taggit
django-ckeditor